refactor: SyncList code generator is reusable for other structs
